Add crude implementation of dictionary command line tool.
Added the basic functionality for the command line tool. This includes
requesting the word from the api, and deserializing the response. Also
the api id, and key are not stored with the repository but are actually
environment variables, this information will need to be added to the
README.md later.